+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix buildbot version in upgraded platforms ([#32](../../pull/32))
 - Fix missing m4 dependencies in Python CLI output ([#33](../../pull/33))
 
+### Changed
+- Optimize build cache for PR builds ([#36](../../pull/36))
+
 ## [2.4.1-1] - 2019-09-20
 ### Added
 - Upgrade to buildbot 2.4.1 ([#29](../../pull/29))